Broken up celery app
